Minor refactoring.

+++ mlmisc/dataset_utils.py
@@ -5,7 +5,7 @@ import torch
 from . import utils as ut
 
 
-def get_class_weights(dataset):
+def get_class_weights(dataset, dtype=torch.float):
   target = np.empty(len(dataset), dtype=np.int32)
   for i in range(len(dataset)):
     x, y = dataset[i]
@@ -14,7 +14,7 @@ def get_class_weights(dataset):
   class_counts = np.unique(target, return_counts=True)[1]
   weight = class_counts / np.sum(class_counts)
 
-  ds_weight = torch.tensor(weight, dtype=torch.float)
+  ds_weight = torch.tensor(weight, dtype=dtype)
 
   alog.debug(f'Dataset class weight: { {c: f"{n * 100:.2f}%" for c, n in enumerate(ds_weight)} }')
